rename dir, temporary drone file updates
